pyproject.toml
[tool.poetry]
name = "metrify"
version = "0.1.0"
description = "Sistema de coleta automática de métricas com base na API do Github."
authors = ["Pedro Binotto <pedrosantibinotto@gmail.com>", "Eric Fernandes Evaristo <ericfernandes7@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pymongo = "^4.7.1"
flask-pymongo = "^2.3.0"
python-dotenv = "^1.0.1"
gql = {extras = ["all"], version = "^3.5.0"}
jwt = "^1.3.1"
flask-apscheduler = "^1.13.1"
sphinx = { version = "^7.3.7", optional = true }
sphinxcontrib-httpdomain = { version = "^1.8.1", optional = true }

[tool.poetry.group.dev.dependencies]  
autopep8 = "^2.1.0"
mypy = "^1.10.0"
pre-commit = "^3.7.0"
pylint = "^3.1.0"
pytest = "^8.2.0"
pytest-sugar = "^1.0.0"
pytest-cov = "^5.0.0"
tox = "^4.15.0"

[tool.poetry.extras]
docs = ["Sphinx", "sphinxcontrib-httpdomain"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-v"
testpaths = ["tests"]

[tool.autopep8]
experimental = true
in-place = true
recursive = true
verbose = true
ignore = "E402"
max_line_length = 120
aggressive = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
